refactor: move all no-code in one component with composables
